Avoid creating large constants in ConvertTFLBroadcastToMulOp optimization pass.
This pattern is inherited from before and has proved to be increasing the model size due the introduction of large splat const. In its current form this pattern replaces a tfl.broadcast_to op (with rank<4) to a tfl.mul with all-ones tensor.

This change will keep the broadcast_to ops as is because its clear that introducing MUL is not an optimization.

PiperOrigin-RevId: 733584588